Authentication Added!
Authentication complete!

+++ Server/routes/users.js
@@ -1,11 +1,13 @@
 const express = require('express')
 const app = express()
+const bcrypt = require('bcrypt')
 const axio = require('axios')
 const User = require("../models/user")
 const Credential = require("../models/credential")
 const _Class = require("../models/class")
 const router = express.Router();
 const uID = require("../JS/UniqueCode")
+const verify = require("./verifyToken")
 
 
 
@@ -22,9 +24,9 @@ router.get("/", async (req , res) => {
 })
 
 //Get all classes that student participated
-router.post("/class", async (req , res) => {
+router.post("/class", verify , async (req , res) => {
     try{
-        const user = await User.findOne({email : req.body.email});
+        const user = await User.findOne({email : req.user.email});
         const classCodes = user.classParticipated;
         const classOwned = user.classOwnerShip;
         var classParticipated = []
@@ -39,7 +41,7 @@ router.post("/class", async (req , res) => {
                 classParticipated.push(nClass)
             }
         }
-
+        console.log({classOwnerShip , classParticipated})
         res.json({classOwnerShip , classParticipated})
         
     }catch(err){
@@ -49,8 +51,15 @@ router.post("/class", async (req , res) => {
     }
 })
 
+//Get user Info
+router.get("/user", verify , async (req, res) => {
+    const email = req.user.email
+    const user = await User.findOne({email})
+    res.json(user)
+})
+
 //Get data for creating class
-router.post("/createclass", async (req, res) => {
+router.post("/createclass", verify , async (req, res) => {
     // Require 2 parameters {classroomName, email}
     try{
         const code = uID(6) // Create a unique keyCode for classroom **Might be duplicated**
@@ -59,10 +68,10 @@ router.post("/createclass", async (req, res) => {
             code: code,
             currentlyStreaming: "",
             members: [],
-            owner: req.body.email
+            owner: req.user.email
         })
         // Update user classOwnerShip and classParticipated
-        const updateUser = await User.updateOne({email:req.body.email}, {$addToSet: { classOwnerShip: code , classParticipated: code} })
+        const updateUser = await User.updateOne({email:req.user.email}, {$addToSet: { classOwnerShip: code , classParticipated: code} })
         if (updateUser){
             const _savedClass = await _class.save()
             res.json(_savedClass)
@@ -78,21 +87,21 @@ router.post("/createclass", async (req, res) => {
     }
 })
 
-router.post("/joinclass", async (req, res) => {
+router.post("/joinclass", verify , async (req, res) => {
     //Require 2 data  {email, code}
    try{
         const _class = await _Class.findOne({code : req.body.code})
         //Check if the class is actually exist
         if (_class){
             // update classParticipated in user's object 
-            const updatedUser = await User.updateOne({email:req.body.email}, {$addToSet: { classParticipated: req.body.code }})
+            const updatedUser = await User.updateOne({email:req.user.email}, {$addToSet: { classParticipated: req.body.code }})
             // Add new member to the class object
-            await _Class.updateOne({code : req.body.code}, {$addToSet: { members: req.body.email }})
+            await _Class.updateOne({code : req.body.code}, {$addToSet: { members: req.user.email }})
             // Check if one array has been added or not, if not, class is already joined!
             if (updatedUser.nModified == 0){
                 res.json({message : "Already join the class!"})
             }else{
-                _class.members.push(req.body.email)
+                _class.members.push(req.user.email)
                 // Send back the class
                 res.json(_class)
             }
@@ -103,49 +112,36 @@ router.post("/joinclass", async (req, res) => {
         }
    }catch(err){
         res.json({
-            message : "Error nob : " + err
+            message : "Error Occured : " + err
         })
    }
 })
 
-router.post("/deleteClass",async (req,res)=>{
+router.post("/deleteClass", verify , async (req,res)=>{
     try{
         const {code} = req.body //pull class code from request body
+        
+        // Check if the user own the class
         const _class = await _Class.findOne({code}) // get the desired class to delete
-        await User.updateOne({email:_class.owner},{$pull:{classOwnerShip:{$in:[code]}}}) //remove ownership
-        await User.updateMany({email:{$in:_class.members}},{$pull:{classParticipated:{$in:[code]}}},{multi:true}) //remove current memebers
-        await _Class.deleteOne({code}) //delete class
-        res.send("success")
+        if (_class.owner == req.user.email){
+            await User.updateOne({email:_class.owner},{$pull:{classOwnerShip:{$in:[code]}}}) //remove ownership
+            await User.updateOne({email:_class.owner},{$pull:{classOwnerShip:{$in:[code]}}}) //remove participation
+            await User.updateMany({email:{$in:_class.members}},{$pull:{classParticipated:{$in:[code]}}},{multi:true}) //remove current memebers
+            await _Class.deleteOne({code}) //delete class
+            res.send({"message" : "Delete class as successfully"})
+        }else{
+            res.send({"message" : "You don't own the class"})
+        }
+
     }catch(err){
-        res.send("error")   
+        res.send({"message" : "Error Occured"})   
     }
 })
 
-//Get Data for sign up
-router.post("/signUp", async (req , res ) => {
-    try{
-        const user = new User({
-            classOwnerShip : [],
-            classParticipated : [],
-            email : req.body.email,
-            name : req.body.name
-        });
-        const credential = new Credential({
-            email : req.body.email,
-            pwd : req.body.pwd
-        })
-        console.log(user)        
-        const savedUser = await user.save();
-        const savedCredential = await credential.save();
-        res.json(savedUser);
-    }catch(err){
 
-    }
-
-})
 
 //Get specific user (Obsolete)
-router.get("/:postId", async (req , res) => {
+router.get("/:postId", verify , async (req , res) => {
     try{
         console.log(req.params.postId)
         const user = await User.findOne({email:req.params.postId});
@@ -157,5 +153,7 @@ router.get("/:postId", async (req , res) => {
     }
 
 })
+
+
 
 module.exports = router;
